Fix type casting for payload value in ElectraProtocol dump function

+++ esphome/components/remote_base/electra_protocol.cpp
@@ -49,7 +49,7 @@ optional<ElectraData> ElectraProtocol::decode(RemoteReceiveData src) {
   return data;
 }
 void ElectraProtocol::dump(const ElectraData &data) {
-  ESP_LOGI(TAG, " payload value: %016X", data.payload.value);
+  ESP_LOGI(TAG, " payload value: %016X", (uint64_t)data.payload.value);
   ESP_LOGI(TAG, "Received Electra: magic=0x%016X, payload=0x%016X", data.magic, data.payload.value);
 
   ESP_LOGI(TAG, "  bytes: %02X %02X %02X %02X %02X %02X %02X %02X", data.payload.bytes[0], data.payload.bytes[1],
